+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove private params from the url. Additional fix [#110](https://github.com/rokwire/safer-illinois-app/issues/110).
 - Add external link icon for SettingsHomePanel buttons [241](https://github.com/rokwire/safer-illinois-app/issues/241).
 - Add building access status to the Home Status widget [#243](https://github.com/rokwire/safer-illinois-app/issues/243)
+- Add building access status to the Home Status widget [#269](https://github.com/rokwire/safer-illinois-app/issues/269)
 
 ## [2.7.6] - 2020-10-13
 ### Changed